Update and rename GelfLayoutAccessLog.java to GelfLayoutAccess.java
